requirements.txt
bitcoinlib==0.6.14
certifi==2024.2.2
charset-normalizer==3.3.2
ecdsa==0.18.0
fastecdsa==2.3.0
greenlet==3.0.3
idna==3.6
numpy==1.26.3
pycryptodome==3.20.0
python-bitcoinrpc==1.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.2.0
